refactor(frontend): make home button group as a independent component, add widget container placeholder
